[vimrc] Set guicursor option
Neovim 0.2.0+ now supports changing the cursor with the 'guicursor'
option.  The only catch is that it has to be reset to a block on
VimLeave because the guicursor setting seems to pass through to the
parent terminal which will have its cursor stuck as a vertical bar on
exiting vim.