[tmux] Fix Gutentags preventing pane switch
Gutentags process contains Vim in the name and was preventing pane
switching until it had finished.

Would usually occur when opening a file in a separate pane for a few
seconds, closing the file, and then trying to switch panes. If the
Gutentags process was still running then tmux would send the
control-chord keys to the pane thinking it was a Vim instance.