[install] Set homebrew bash as default shell
Related to recent tmux changes. I noticed that a zsh shell was being
started even though I have my `default_command` in tmux set to invoke
bash. Worried that this will cause problems or be inefficient since it
has to start 2 processes for each shell. Just set the default which is
probably a better practice anyway.